requirements.txt
alembic==1.8.1
asn1crypto==1.5.1
attrs==22.1.0
autopage==0.5.1
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cliff==4.1.0
cloudpickle==2.0.0
cmaes==0.9.0
cmd2==2.4.2
colorlog==6.7.0
contourpy==1.0.6
cramjam==2.6.2
cryptography==38.0.4
cycler==0.11.0
fastparquet==2022.12.0
filelock==3.8.2
fonttools==4.38.0
fsspec==2022.11.0
greenlet==2.0.1
idna==3.4
imbalanced-learn==0.9.1
importlib-metadata==4.13.0
importlib-resources==5.10.1
joblib==1.2.0
kiwisolver==1.4.4
lightgbm==3.3.3
Mako==1.2.4
MarkupSafe==2.1.1
matplotlib==3.6.2
numpy==1.23.5
optuna==3.0.4
oscrypto==1.3.0
packaging==21.3
pandas==1.5.2
pbr==5.11.0
Pillow==9.3.0
prettytable==3.5.0
pyarrow==8.0.0
pycparser==2.21
pycryptodomex==3.16.0
PyJWT==2.6.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyperclip==1.8.2
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
requests==2.28.1
scikit-learn==1.1.3
scipy==1.8.1
seaborn==0.12.1
six==1.16.0
snowflake-connector-python==2.8.3
snowflake-snowpark-python==1.0.0
SQLAlchemy==1.4.44
stevedore==4.1.1
threadpoolctl==3.1.0
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.5
xgboost==1.7.1
zipp==3.11.0
